+++ data_mgmt/2_catchments/aggregate_basins.py
@@ -659,7 +659,7 @@ if __name__ == "__main__":
         ,55,56,57,61,62,63,64,65,66,67,71,72,73,74,75,76,77,78,81,82,83,84,85,86, 54]
 
     for BASIN in basins:
-        DB_PATH = f"C:/Users/mheberger/AppData/Local/delineator/v2.2/basins{BASIN}.db"
+        DB_PATH = f"C:/Users/mheberger/AppData/Local/delineator/basins{BASIN}.db"
         CAT_SHP = rf"C:\Data\GIS\MERITBasins\catchments\bugfix_repaired\cat_pfaf_{BASIN}_MERIT_Hydro_v07_Basins_v01_bugfix1.shp"
         RIV_SHP = rf"C:\Data\GIS\MERITBasins\rivers\bugfix_repaired\riv_pfaf_{BASIN}_MERIT_Hydro_v07_Basins_v01_bugfix1.shp"